Refactor extras forms
