+++ netbox/extras/forms/__init__.py
@@ -1,0 +1,6 @@
+from .models import *
+from .filtersets import *
+from .bulk_edit import *
+from .bulk_import import *
+from .customfields import *
+from .scripts import *
